+++ src/presentation/controllers/login/login.spec.ts
@@ -1,8 +1,7 @@
-import { InvalidParamError, MissingParamError } from '../../errors'
 import { LoginController } from './login'
+import { InvalidParamError, MissingParamError } from '../../errors'
+import { EmailValidator, HttpRequest, Authentication } from './login-protocols'
 import { badRequest, serverError, unauthorized } from '../../helpers/http-helper'
-import { EmailValidator, HttpRequest } from '../signup/signup-protocols'
-import { Authentication } from '../../../domain/usecases/authentication'
 
 interface SutTypes {
   sut: LoginController
